Fix pipeline navigation header
The navigation header was incorrectly
being rendered as part of the body,
resulting in a large white margin. Applied
no-body to the card. Opted for this
over slots as more minimal change.

Also, we weren't applying an active
style to the selected tab. This adds
one, making it easier to see what's
active.