Allow well id to be passed into events
Rather than using getWellAt in the template
for loop, we instead construct the merged
object in a computed propeties, and loop
over that instead,

This makes well.id available in the template.

I also move plateMap out of the data and just grab it from the store
directly, no need to duplicate it all over the place.

Did consider something similar for the well data. It feels more 'proper'
to me, but does have implications wrt.
reactivity.