Correctly handle promises from update
I really wanted to destructure my promise here:
```
const promise = request.update(payload({ libraries, pool }), { include: 'tube' })
```

But while it seemed to work in reality, the mocked version in the
tests wasn't behaving the same. I'd look in to it more, but this
behaviour is due to change very shortly.