Add darag-select to plates
This uses the library 'vue-selecto' which appears to be one of the better
support js drag-selection libraries. That said, vue integration isn't
fantastic, and we end up dealing with events that emit raw SVG elements,
rather than vue components.

There was also a bit of trouble limiting the selection with drag dragContainer
where the component appeared to to ignore the configuration.