Centralise alert handling
This migration centralises the alert handling through the store,
avoiding the need for alerts within each of the components. This
makes it substantially easier to handle errors within store actions
and commits.

I've tried to reduce the level of disruption caused by these
changes, and still need to ensure that the dismissal of a message
properly removes it from the array.