Remove moment
Moment is a date formatting library, which was solely used
for formatting the date on labels. However, the library home page itself
states:

The Moment team has discussed these issues at length. We recognize that many existing projects may continue to use Moment, but we would like to discourage Moment from being used in new projects going forward. Instead, we would like to recommend alternatives that are excellent choices for use in modern applications today. We would also like to promote the Temporal addition to the JavaScript language, which is looking for feedback and contributors.

Hwever, it seems unnecessary to import an whole library for
what can be accomplished in a couple of lines of code.