Import requests rather than plates
In order to allow the import of tubes, as well as plates, we switch
to importing via request. Currently the API doesn't support
including tubes with the requests, so we may need to do a bit of work
there. I suspect we *do* have sufficient information included on
request itself, but things may be tidier overall if we actually import
the tubes.