Don't render tag list by default
We pop the tag list in a details section to allow it to be shown
only when needed. This prevents the pool setup from being forced
too far down the page. We add a tag count badge to reassure users
that their tag set has correctly loaded.

This change is made of the assumption that the tag information in
the select dropdowns is sufficient the majority of the time. It may
be that UAT results in this change rolling back.