Convert alerts to toasts
Toasts are notifications rendered relative to the viewport,
in this case the bottom right. This ensures they are always visible,
regardless of the position of the viewport.

This is achieved by wrapping the messages in a message container div.
This div is fixed relative the the viewport, and is positioned in the
bottom right. We have a high z-index to position us above any modals,
as, for better or worse, we use alerts to provide modal feedback.