Update tailwind
Our updating of the rest of the toolchain means we no longer
rely on older versions of post-css, so can update tailwind.

I've avoided the leap to v3 for the moment, as don't want
to deal with the breaking changes.

This also adds the programatically built index.css to the
git-ignore, to avoid excessive churn. That file gets *big*
and doesn't need to be in the repo.