Add FlaggedFeature Component
I'm currently having trouble getting the
tests to fuction as intended. While the
code seems to function as intended in reality
I'm getting the following error when running tests:

Error: Could not get current instance, check to make sure that `useSwrv` is declared in the top level of the setup function.

Investigating