Add basic loading page
With is static at the moment, it would be nice to add a little
bit of animation, but keeping things simple for the moment.

Note: We probably should be using 'object' rather than 'img' to
load the SVG, as this includes it with the document, and allows
us to, say, adapt the styles. However that also requires us
to tweak the SVG slightly.