Fix other http warnings
This is going to be a case of wack a mole.

We've got a couple of options:

- explicitly stub out the vuex components in the unit tests. This will
  probably mean getting rid of the global localVue and ensure we
  instantiate it per test.
- Use something like mirage to set up the basic HTTP responses globably.

While the former firs best with a strict unit testing approach, I've
found the latter is actually less error prone, and would probably
simplify a lot of the code elsewhere. That said, I don't think approach
one is without value, and might be useful in some contexts.