refactor: Update printers getter to filter by labwareType
The printers getter in printing.js has been updated to include a labwareType parameter. This allows for filtering the printers based on the specified labware type. If a labwareType is provided, the getter will return an array of printers that match the labwareType. If no labwareType is provided, it will return all printers.